Moving class LocationManager to its own file
